Feat: Refactor plugin to use factory-singleton pattern
